+++ src/Generators/SphereParticleGenerator.h
@@ -19,7 +19,7 @@ private:
     // Radius r of the Sphere
      */
 
-    std::array<int, 3> x;
+    std::array<double, 3> x;
     double h;
     double m;
     std::array<double, 3> v;
@@ -31,7 +31,7 @@ public:
      * create a N1xN2xN3 grid of particles with mass m, initial velocity v, left corner coordinate (x,y,z)
     */
 
-    SphereParticleGenerator(std::array<int, 3> x1, double h1, double m1, std::array<double, 3> v1, double r1, int type1);
+    SphereParticleGenerator(std::array<double, 3> x1, double h1, double m1, std::array<double, 3> v1, double r1, int type1);
 
     void generateParticles(ParticleContainerBase& particleContainer);
 };
